LayoutTests/css3/filters/custom-with-at-rule-syntax/parsing-src-property-invalid-expected.txt

Test src property parsing in the @-webkit-filter at-rule.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



========================================
URL tests.
========================================

No property value.
src: ;
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('src') is null

Only a single comma.
src: ,;
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('src') is null

Only two commas.
src: , ,;
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('src') is null

Identifier instead of URL.
src: vertex-shader;
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('src') is null

Identifier with format instead of URL with format.
src: vertex-shader format('x-shader/x-vertex');
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('src') is null

Local function instead of URL function.
src: local(shader.vs);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('src') is null

========================================
Format function tests.
========================================

Identifier instead of format function.
src: url(shader.vs) vertex-shader;
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('src') is null

String instead of format function.
src: url(shader.vs) 'x-shader/x-vertex';
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('src') is null

Format function misspelled.
src: url(shader.vs) frmt('x-shader/x-vertex');
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('src') is null

Format function empty.
src: url(shader.vs) format();
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('src') is null

Format function with single identifer argument.
src: url(shader.vs) format(vertex-shader);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('src') is null

Format function with single number argument.
src: url(shader.vs) format(0);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('src') is null

Format function with multiple space-separated identifer arguments.
src: url(shader.vs) format(vertex-shader fragment-shader);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('src') is null

Format function with multiple comma-separated identifer arguments.
src: url(shader.vs) format(vertex-shader,fragment-shader);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('src') is null

Format function with multiple space-separated string arguments.
src: url(shader.vs) format('x-shader/x-vertex' 'x-shader/x-fragment');
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('src') is null

Format function with multiple comma-separated string arguments.
src: url(shader.vs) format('x-shader/x-vertex','x-shader/x-fragment');
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('src') is null

Multiple valid format functions.
src: url(shader.vs) format('x-shader/x-vertex') format('x-shader/x-fragment');
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('src') is null

========================================
Punctuation tests.
========================================

Leading comma before first URL.
src: , url(shader.vs) format('x-shader/x-vertex');
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('src') is null

Comma between URL and format.
src: url(shader.vs), format('x-shader/x-vertex');
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('src') is null

Trailing comma after format.
src: url(shader.vs) format('x-shader/x-vertex'),;
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('src') is null

Two commas between URLs.
src: url(shader.vs),,url(shader.fs);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('src') is null

Two commas between URLs with format.
src: url(shader.vs) format('x-shader/x-vertex'),,url(shader.fs) format('x-shader/x-fragment');
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('src') is null

No comma between URLs.
src: url(shader.vs) url(shader.fs);
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('src') is null

No comma between URLs with format.
src: url(shader.vs) format('x-shader/x-vertex') url(shader.fs) format('x-shader/x-fragment');
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('src') is null
PASS successfullyParsed is true

TEST COMPLETE
